Add: footer 번역기 컴포넌트화 및  map사용
