+++ src/Pages/Account/components/AccountFooter/AccountFooter.scss
@@ -1,30 +1,15 @@
-@import "../../../styles/common.scss";
+@import "../../../../styles/common.scss";
 
 .accountfooter {
-  //border: 1px solid orange;
   height: 38px;
   width: 100%;
+  display: flex;
+  justify-content: space-between;
   position: relative;
   padding: 0 14px;
   margin-top: 50px;
 
-  select {
-    position: absolute;
-    left: 14px;
-    border: 0px;
-    font-family: $loginFontFamily;
-
-    &:focus {
-      outline: none;
-      border-bottom: 3px solid #1087dd;
-    }
-    option {
-      color: black;
-    }
-  }
-
   div {
-    position: absolute;
     right: 14px;
     padding-left: 0px;
     span {
