+++ src/Pages/Account/components/AccountFooter/Languagedata.js
@@ -1,0 +1,16 @@
+const FOOTERLANGUAGE = [
+  "ENGLISH",
+  "DANISH",
+  "GERMAN",
+  "SPANISH",
+  "FRENCH",
+  "ITALIAN",
+  "JAPANESE",
+  "KOREAN",
+  "DUTCH",
+  "RUSSIAN",
+  "CHINESE",
+  "CHINESE TRADITIONAL",
+];
+
+export default FOOTERLANGUAGE;
